Music Button and SFX Button Update
Null hatası giderildi mute edildiğinde ilgili görsel değiştirildi değer 0 olursa ilgili görsel yine değiştirildi

+++ Assets/Project/Scripts/Towers/SniperTower/TowerTarget.cs
@@ -24,7 +24,7 @@ public class TowerTarget : MonoBehaviour
     public Transform[] muzzlePosition;
     [Header("Tower Bullet Points")]
     public Transform[] firePoints;
-    public int firePointIndex = 0;
+    public sbyte firePointIndex = 0;
     public float recoil;
     public float RecoilSpeed;
     public GameObject Muzzle;
